First commit
A Python project which returns the Meaining,translation, antonym & synonym of a word by choice of the user.It can also return the meaning of the Text fetched instantly from wikipedia. It returns the result in Text as well as Audio format.

This project is developed by Shubhankar Tiwary (E54).

For the 2nd Assignment of the subject Python with Django(3rd Sem).